requirements.txt
streamlit
torch
torchvision
matplotlib
seaborn
numpy
Pillow
scikit-learn
kagglehub
